+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 ## 8.15.13 (Unreleased)
 
-### Features
-
- - Add Input Action: Execute Python 3 Code ([#1334](https://github.com/kizniche/Mycodo/issues/1334))
-
 ### Bugfixes
 
+ - Fix inability to properly import settings backup
  - Fix Actions not executing for MQTT, TTN, and Python Code Inputs ([#1336](https://github.com/kizniche/Mycodo/issues/1336))
- - Fix Python Code Inputs running pylint
+ - Fix running pylint
+
+### Features
+
+ - Add ability to use Actions in Custom Functions
+ - Add Input Action: Execute Python 3 Code ([#1334](https://github.com/kizniche/Mycodo/issues/1334))
+ - Add Function: Adafruit Neokey (Key Press Executes Actions) ([#1353](https://github.com/kizniche/Mycodo/issues/1353))
+ - Add Action: Neopixel Flashing On
+ - Add Action: Neopixel Flashing Off
+ - Change deprecated threading.currentThread to threading.current_thread
+
+### Miscellaneous
+
+ - Update InfluxDB 2.x to 2.7.3
 
 
 ## 8.15.12 (2023.10.28)